COM-2989: move function in helper
