+++ src/screens/courses/profile/helper.tsx
@@ -1,8 +1,10 @@
 import { View } from 'react-native';
+import get from 'lodash/get';
 import { E_LEARNING, ON_SITE, REMOTE } from '../../../core/data/constants';
 import LiveCell from '../../../components/steps/LiveCell';
 import ELearningCell from '../../../components/ELearningCell';
 import styles from './styles';
+import { BlendedCourseType } from '../../../types/CourseTypes';
 
 export const renderStepCell = ({ item, index }, course, route) => {
   if ([ON_SITE, REMOTE].includes(item.type)) {
@@ -18,3 +20,12 @@ export const renderStepCell = ({ item, index }, course, route) => {
 };
 
 export const renderSeparator = () => <View style={styles.separator} />;
+
+export const getTitle = (course) => {
+  if (!course) return '';
+  const programName = get(course, 'subProgram.program.name') || '';
+  if (course?.subProgram.isStrictlyELearning) return programName;
+
+  const { misc } = (course as BlendedCourseType);
+  return `${programName}${misc ? `- ${misc}` : ''}`;
+};
